Revert "ftrace import: fix timestamp conversion"
This reverts commit 1e2cf85db31207f86c3639b50d8d11f63ce85213.

+++ timechart/backends/ftrace.py
@@ -1,4 +1,3 @@
-from decimal import *
 import re
 import sys,os
 from timechart.plugin import get_plugins_additional_ftrace_parsers
@@ -167,7 +166,7 @@ def parse_ftrace(filename,callback):
                 'common_comm' : groups[0],
                 'common_pid' :  int(groups[1]),
                 'common_cpu' : int(groups[2]),
-                'timestamp' : int(Decimal(groups[3])*1000000),
+                'timestamp' : int(float(groups[3])*1000000),
                 'event' : event_name,
                 'event_arg' : groups[5]
                 }
